feat(config,engine,mcp): build the intent-scope and sequence-gate designs (59 of 59)
MODEL-SHAPE-1/2 were the last two kernels without a differential binding.
Both modelled a design production did not implement, so binding them needed a
decision first: reshape the kernels to production, or build the design. This
builds the design. Doing so surfaced SCOPE-NOOP-1.

SCOPE-NOOP-1 — taint-driven scope narrowing did nothing.

IntentScopeConfig::restrict_to_trust_floor narrowed scope by filtering
allowed_sink_classes. An empty list there means "unrestricted", so filtering it
returned another empty list, and restricting the default configuration narrowed
nothing:

    IntentScopeConfig::default()
        .restrict_to_trust_floor(TrustTier::Quarantined)
        .check_in_scope("t", SinkClass::PolicyMutation)   // => InScope

PolicyMutation, the highest-privilege sink, survived a restriction to
Quarantined, the lowest trust floor. The kernel has proved SCOPE-1 and SCOPE-2 —
restriction is a subset, scope only narrows — since Phase 6.2E, and both were
false in production for the default config.

Two further holes are closed with it. Both relay call sites computed the
restricted scope and discarded it (`let _ = restricted; // full enforcement in
6.2C`), so narrowing never persisted. And check_in_scope had no caller anywhere
in the relay. Nothing in the workspace's tests caught any of this.

MODEL-SHAPE-1. vellaveto-types::verified_intent_scope::ScopeMask makes scope a
bitmask, so restriction is the bitwise AND the kernel proves and "allow nothing"
and "allow everything" are distinct values. The config surface keeps
allowed_sink_classes; the empty-means-everything convention is converted at
exactly one place. The mask is u16 because SinkClass has nine variants — the
kernel's u8 could not represent rank 8 at all, the same root cause as
TAINT-MODEL-DRIFT. Kernel widened to match; re-verified 16/0.

MODEL-SHAPE-2. vellaveto-engine::verified_sequence_gate is the restriction gate
the kernel modelled, with named RESTRICTION_THRESHOLD and WARMUP_CALLS. It
drives SequenceConfig::default() and decides at the relay call site that
previously held a bare 70 whose result was thrown away. Confidence is u32 on
both sides now; the kernel's u8 was an arbitrary narrowing. Re-verified 22/0.

Enforcement. RelayState carries session_scope, so successive narrowings compose
from the scope in force rather than from the immutable per-relay config. The
scope is consulted on the tool-call path before forwarding, honouring
out_of_scope_action. Deny and RequireApproval both refuse: the stdio relay has
no interactive approval channel there, so RequireApproval is honoured in the
fail-closed direction rather than degraded to an allow. The sink class is
resolved by config, then by the annotation-aware inference already in
helpers.rs, then read-only — without that middle step an unclassified
execute_shell would resolve to ReadOnly and pass a read-only scope.

This tightens real behaviour. The vault and sandworm-hardened presets set
allowed_sink_classes with out_of_scope_action = Deny; that was inert and now
enforces. Regression tests pin that an unconfigured scope still admits every
sink class, that a high trust floor narrows nothing, that narrowing is graded
rather than all-or-nothing, and that a narrowed scope survives a serde round
trip.

GUARD-COMMENT-1 — two of the seven new parity checks could not fail.

Found by mutation-testing them rather than trusting that a passing check means
anything. One matched the relay's own explanatory comment rather than the call;
the other matched a renamed function by prefix. Both now require a call or
definition and are anchored so a line comment cannot satisfy them. Eight cases
covering these checks are in guard-selftest.sh so the property is re-tested on
every run instead of established once by hand.

Also fixes two CI failures unrelated to this work: a yanked chacha20 0.10.1
(cargo update to 0.10.2, cargo-deny advisories ok) and stale evidence blocks in
README.md, docs/ASSURANCE_CASE.md and formal/README.md.

Measured trusted base: 59 of 59 (47 discharged + 5 partial + 7 property).

Gates: fmt, workspace clippy under -Dwarnings, verus parity, trusted
assumptions, kani parity, evidence-check, and 11,858 workspace tests — 0
failures.

+++ formal/verus/verified_sequence_analysis.rs
@@ -37,7 +37,7 @@ verus! {
 pub struct SequenceState {
     pub call_count: u32,
     pub anomaly_detected: bool,
-    pub anomaly_confidence: u8,   // 0-100
+    pub anomaly_confidence: u32,  // 0-100; u32 to match SequenceAnomaly.confidence
     pub taint_active: bool,
     pub new_tools_after_taint: u32,
     pub scope_restricted: bool,
@@ -45,8 +45,8 @@ pub struct SequenceState {
 
 pub const WARMUP_CALLS: u32 = 3;
 pub const MAX_NEW_TOOLS: u32 = 2;
-pub const RESTRICTION_THRESHOLD: u8 = 70;
-pub const MAX_CONFIDENCE: u8 = 100;
+pub const RESTRICTION_THRESHOLD: u32 = 70;
+pub const MAX_CONFIDENCE: u32 = 100;
 
 // ── Core spec functions ────────────────────────────────────────────
 
@@ -60,16 +60,16 @@ pub open spec fn spec_update_anomaly_detected(
 
 /// Spec: update confidence (max-tracking — only increases).
 pub open spec fn spec_update_confidence(
-    current: u8,
-    new_confidence: u8,
-) -> u8 {
+    current: u32,
+    new_confidence: u32,
+) -> u32 {
     if new_confidence > current { new_confidence } else { current }
 }
 
 /// Spec: should scope be restricted?
 pub open spec fn spec_should_restrict(
     anomaly_detected: bool,
-    anomaly_confidence: u8,
+    anomaly_confidence: u32,
 ) -> bool {
     anomaly_detected && anomaly_confidence >= RESTRICTION_THRESHOLD
 }
@@ -111,7 +111,7 @@ pub proof fn lemma_anomaly_monotonic_chain(
 
 // ── SEQ-2: Confidence only increases ──────────────────────────────
 
-pub fn update_confidence(current: u8, new_confidence: u8) -> (result: u8)
+pub fn update_confidence(current: u32, new_confidence: u32) -> (result: u32)
     ensures
         result == spec_update_confidence(current, new_confidence),
         result >= current,    // Never decreases
@@ -120,13 +120,13 @@ pub fn update_confidence(current: u8, new_confidence: u8) -> (result: u8)
     if new_confidence > current { new_confidence } else { current }
 }
 
-pub proof fn lemma_confidence_monotonic(current: u8, new_confidence: u8)
+pub proof fn lemma_confidence_monotonic(current: u32, new_confidence: u32)
     ensures
         spec_update_confidence(current, new_confidence) >= current,
 {
 }
 
-pub proof fn lemma_confidence_chain_monotonic(c0: u8, n1: u8, n2: u8)
+pub proof fn lemma_confidence_chain_monotonic(c0: u32, n1: u32, n2: u32)
     ensures ({
         let c1 = spec_update_confidence(c0, n1);
         let c2 = spec_update_confidence(c1, n2);
@@ -139,7 +139,7 @@ pub proof fn lemma_confidence_chain_monotonic(c0: u8, n1: u8, n2: u8)
 
 pub proof fn lemma_restriction_requires_anomaly(
     anomaly_detected: bool,
-    confidence: u8,
+    confidence: u32,
 )
     ensures
         spec_should_restrict(anomaly_detected, confidence) ==> anomaly_detected,
@@ -147,7 +147,7 @@ pub proof fn lemma_restriction_requires_anomaly(
 {
 }
 
-pub proof fn lemma_no_anomaly_no_restriction(confidence: u8)
+pub proof fn lemma_no_anomaly_no_restriction(confidence: u32)
     ensures
         !spec_should_restrict(false, confidence),
 {
@@ -214,7 +214,7 @@ pub proof fn lemma_high_confidence_triggers_restriction()
 }
 
 /// Composition: detection at confidence >= 70 always restricts.
-pub proof fn lemma_detection_above_threshold_restricts(confidence: u8)
+pub proof fn lemma_detection_above_threshold_restricts(confidence: u32)
     requires
         confidence >= RESTRICTION_THRESHOLD,
     ensures
@@ -229,7 +229,7 @@ pub proof fn lemma_detection_above_threshold_restricts(confidence: u8)
 pub fn sequence_step(
     state: &SequenceState,
     detected: bool,
-    confidence: u8,
+    confidence: u32,
     is_novel_after_taint: bool,
 ) -> (new_state: SequenceState)
     requires
